add unenroll and remove from favorite

+++ src/main/java/individualassignment/edubridge/business/users/student/impl/DeleteStudentUseCaseImpl.java
@@ -1,6 +1,7 @@
 package individualassignment.edubridge.business.users.student.impl;
 
 import individualassignment.edubridge.business.users.student.DeleteStudentUseCase;
+import individualassignment.edubridge.business.users.student.exceptions.InvalidStudentException;
 import individualassignment.edubridge.persistence.users.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,8 @@ public class DeleteStudentUseCaseImpl implements DeleteStudentUseCase {
 
     @Override
     public void deleteStudent(long studentId) {
+        if(!studentRepository.existsById(studentId))
+            throw new InvalidStudentException();
         studentRepository.deleteById(studentId);
     }
 }
